c2 server/requirements.txt
Django>=3.0,<4.0
psycopg2-binary>=2.8
djangorestframework
pytest-django
requests
